# Reuse frontmatter parser in classifier