While looking at #4691 , I noticed that the code in `classifier.go` does some ad-hoc parsing for HTML vs other content.

Given that a page without frontmatter can't be usefully parsed, it seemed more straightforward to reuse the frontmatter parser to determine whether the HTML file was content or a plain file.

Unfortunately, it turns out that TestSectionNaming has some HTML documents that are just strings without frontmatter, so I ended up needing to adjust those tests.

I'm wondering whether it makes sense to apply this "has frontmatter" test universally in ClassifyContentFile; it appears that doing so would need updates in the following tests from hugolib:

    TestCascade
    TestContentMap
    TestRenderStringOnListPage
    TestSiteBuildErrors
    TestShortCodesInSite
    TestSectionNaming
    TestRefLinking
    TestCrossrefs
    TestDataFromShortcode