@bep 2020-03-24
Thanks for this, and while I  agree that this in general looks like a good thing, one quick note: The reason I implemented this buffered "quick detector" for pure HTML files was to avoid having to read the entire file into memory.

We have some benchmarks that usually shows these issues named "BenchmarkSiteNew/*" (or something).

One tip would to have a look at this tool: https://github.com/bep/gobench

Not much documentation, but look at the `hugobench -h` output.