@bep 2020-03-24
Note that I don't think we use HTML files in the above mentioned benchmarks, though.