@bep 2020-03-24
Or, as it is for HTML only, maybe it doesn't matter too much the memory overhead...