@evankanderson 2020-03-24
I changed BenchmarkSiteNew to use HTML files for most of the content instead of markdown.

The result is around a 8% increase in allocations and somewhere between 1% and 17% on CPU time (17% for "edit" case, 1% for "regular" case).

I can work on splitting `ParseFrontMatterAndContent` into a `ParseFrontMatter` which reads from the file more carefully until `Iterator.Input()` is called.

I'm also in the process of rerunning the benchmark to see if I can get more stable results; I suspect I'm thermal throttling on my laptop so I may have to switch to another machine.

```
 go version go1.13.5 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/hugolib
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      30	  42251521 ns/op	26922799 B/op	  255512 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      33	  34802199 ns/op	26947088 B/op	  255513 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      34	  34643581 ns/op	26916856 B/op	  255498 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      32	  33918035 ns/op	26933820 B/op	  255511 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  48963396 ns/op	32867477 B/op	  320103 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  50271350 ns/op	32937488 B/op	  320113 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  51166796 ns/op	32900642 B/op	  320124 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  61399388 ns/op	32884646 B/op	  320115 allocs/op
PASS
ok  	github.com/gohugoio/hugo/hugolib	21.834s
Switched to branch 'classier'
Your branch and 'upstream/master' have diverged,
and have 3 and 1 different commits each, respectively.
  (use "git pull" to merge the remote branch into yours)


 go version go1.13.5 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/hugolib
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      22	  48980276 ns/op	28580883 B/op	  275674 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      28	  38236375 ns/op	28590353 B/op	  275676 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      30	  40761404 ns/op	28575465 B/op	  275661 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      31	  39999467 ns/op	28566667 B/op	  275663 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  54808500 ns/op	34561612 B/op	  340271 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  54804296 ns/op	34575206 B/op	  340281 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  55202866 ns/op	34568818 B/op	  340272 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  54575961 ns/op	34559677 B/op	  340273 allocs/op
PASS
ok  	github.com/gohugoio/hugo/hugolib	17.227s


name                                 old time/op    new time/op    delta
SiteNew/Edit_Deep_content_tree-8       36.4ms ±16%    42.0ms ±17%    ~     (p=0.200 n=4+4)
SiteNew/Regular_Deep_content_tree-8    53.0ms ±16%    54.8ms ± 1%    ~     (p=0.343 n=4+4)

name                                 old alloc/op   new alloc/op   delta
SiteNew/Edit_Deep_content_tree-8       26.9MB ± 0%    28.6MB ± 0%  +6.12%  (p=0.029 n=4+4)
SiteNew/Regular_Deep_content_tree-8    32.9MB ± 0%    34.6MB ± 0%  +5.07%  (p=0.029 n=4+4)

name                                 old allocs/op  new allocs/op  delta
SiteNew/Edit_Deep_content_tree-8         256k ± 0%      276k ± 0%  +7.89%  (p=0.029 n=4+4)
SiteNew/Regular_Deep_content_tree-8      320k ± 0%      340k ± 0%  +6.30%  (p=0.029 n=4+4)
```