@evankanderson 2020-03-24
Here are --count=10 results, which show around a 6% increase in CPU time for the "edit" case and 0% for the "regular" case, along with the same memory allocation deltas as before.

```
Benchmark and compare branch "clean" and "classier".
Switched to branch 'clean'
Your branch and 'upstream/master' have diverged,
and have 1 and 1 different commits each, respectively.
  (use "git pull" to merge the remote branch into yours)


 go version go1.13.5 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/hugolib
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      31	  38821229 ns/op	26933989 B/op	  255510 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      31	  34190873 ns/op	26924626 B/op	  255505 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      27	  37272635 ns/op	26933585 B/op	  255508 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      32	  36690488 ns/op	26936151 B/op	  255506 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      32	  36147433 ns/op	26927754 B/op	  255517 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      30	  37170453 ns/op	26934790 B/op	  255517 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      32	  36515062 ns/op	26930886 B/op	  255515 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      32	  39681874 ns/op	26935943 B/op	  255515 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      31	  40256163 ns/op	26929189 B/op	  255510 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      31	  42475208 ns/op	26937511 B/op	  255514 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  51928399 ns/op	32880291 B/op	  320115 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  59874902 ns/op	32886961 B/op	  320116 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  53097418 ns/op	32931261 B/op	  320117 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  58198183 ns/op	32929927 B/op	  320119 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  60903281 ns/op	32950847 B/op	  320113 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      21	  52178128 ns/op	32950346 B/op	  320115 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      21	  54204179 ns/op	32889567 B/op	  320099 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  54095487 ns/op	32922389 B/op	  320122 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      21	  54844351 ns/op	32923675 B/op	  320120 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      22	  52709922 ns/op	32961530 B/op	  320119 allocs/op
PASS
ok  	github.com/gohugoio/hugo/hugolib	53.162s
Switched to branch 'classier'
Your branch and 'upstream/master' have diverged,
and have 3 and 1 different commits each, respectively.
  (use "git pull" to merge the remote branch into yours)


 go version go1.13.5 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/hugolib
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      26	  40784890 ns/op	28583137 B/op	  275675 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      33	  40689480 ns/op	28577316 B/op	  275674 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      31	  42765048 ns/op	28566673 B/op	  275664 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      28	  40516672 ns/op	28575105 B/op	  275666 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      28	  39034358 ns/op	28574294 B/op	  275664 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      25	  40743277 ns/op	28570709 B/op	  275675 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      26	  40784004 ns/op	28589587 B/op	  275674 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      27	  38129513 ns/op	28566417 B/op	  275663 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      30	  42182210 ns/op	28573935 B/op	  275665 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      28	  38228966 ns/op	28592859 B/op	  275670 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      21	  54590331 ns/op	34566702 B/op	  340271 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  54788696 ns/op	34591464 B/op	  340278 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  55449936 ns/op	34562114 B/op	  340279 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  58160737 ns/op	34541733 B/op	  340265 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  57530338 ns/op	34624665 B/op	  340277 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  54563267 ns/op	34647628 B/op	  340284 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      20	  52997614 ns/op	34585548 B/op	  340276 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  54799260 ns/op	34552685 B/op	  340272 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  57230261 ns/op	34590016 B/op	  340271 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      19	  57058014 ns/op	34571560 B/op	  340278 allocs/op
PASS
ok  	github.com/gohugoio/hugo/hugolib	43.979s


name                                 old time/op    new time/op    delta
SiteNew/Edit_Deep_content_tree-8       37.9ms ±12%    40.4ms ± 6%  +6.50%  (p=0.011 n=10+10)
SiteNew/Regular_Deep_content_tree-8    55.2ms ±10%    55.7ms ± 5%    ~     (p=0.393 n=10+10)

name                                 old alloc/op   new alloc/op   delta
SiteNew/Edit_Deep_content_tree-8       26.9MB ± 0%    28.6MB ± 0%  +6.11%  (p=0.000 n=10+10)
SiteNew/Regular_Deep_content_tree-8    32.9MB ± 0%    34.6MB ± 0%  +5.02%  (p=0.000 n=10+9)

name                                 old allocs/op  new allocs/op  delta
SiteNew/Edit_Deep_content_tree-8         256k ± 0%      276k ± 0%  +7.89%  (p=0.000 n=10+10)
SiteNew/Regular_Deep_content_tree-8      320k ± 0%      340k ± 0%  +6.30%  (p=0.000 n=9+10)
```