@evankanderson 2020-03-25
Trying again with larger HTML does reveal a 10-15% difference in CPU time with much larger HTML files (which is probably more realistic -- see https://github.com/gohugoio/hugo/pull/7088/commits/95b9ef9b510ecb3668362ade97cd00e4313e84a4 and https://github.com/gohugoio/hugo/pull/7088/commits/9c034df72f83eccda3979b8f94527fd258424737 for the changes.

Would you like me to commit at least https://github.com/gohugoio/hugo/pull/7088/commits/95b9ef9b510ecb3668362ade97cd00e4313e84a4 (separately) to have HTML files in the benchmark?

```
$ gobench -package=./hugolib -bench="BenchmarkSiteNew/Deep_content_tree" -base clean -count 10
Benchmark and compare branch "clean" and "classier".
Switched to branch 'clean'
Your branch and 'upstream/master' have diverged,
and have 2 and 1 different commits each, respectively.
  (use "git pull" to merge the remote branch into yours)


 go version go1.13.5 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/hugolib
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  98541868 ns/op	50233599 B/op	  349514 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      15	  74769839 ns/op	50239447 B/op	  349495 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  78039988 ns/op	50231674 B/op	  349492 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  79725531 ns/op	50256285 B/op	  349498 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      15	  70621619 ns/op	50208484 B/op	  349505 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      16	  78064357 ns/op	50231592 B/op	  349487 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      18	  82328296 ns/op	50194910 B/op	  349490 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  80069058 ns/op	50248842 B/op	  349496 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  73815900 ns/op	50216518 B/op	  349504 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  81175073 ns/op	50241414 B/op	  349500 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  98702417 ns/op	64321724 B/op	  415436 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  93257931 ns/op	64351320 B/op	  415418 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  96247710 ns/op	64340197 B/op	  415418 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  91141394 ns/op	64289131 B/op	  415409 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  92614521 ns/op	64348725 B/op	  415436 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  93837764 ns/op	64306518 B/op	  415414 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  95883651 ns/op	64349421 B/op	  415415 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  96928766 ns/op	64345728 B/op	  415432 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  92704936 ns/op	64307549 B/op	  415431 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      12	  90979902 ns/op	64307488 B/op	  415427 allocs/op
PASS
ok  	github.com/gohugoio/hugo/hugolib	57.664s
Switched to branch 'classier'
Your branch and 'upstream/master' have diverged,
and have 4 and 1 different commits each, respectively.
  (use "git pull" to merge the remote branch into yours)


 go version go1.13.5 darwin/amd64

goos: darwin
goarch: amd64
pkg: github.com/gohugoio/hugo/hugolib
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      12	 100225984 ns/op	62714252 B/op	  384514 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      14	  83270607 ns/op	62709312 B/op	  384496 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  80275202 ns/op	62742745 B/op	  384495 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  90625646 ns/op	62719070 B/op	  384504 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  83803064 ns/op	62704004 B/op	  384490 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  79374122 ns/op	62756000 B/op	  384498 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  77768385 ns/op	62738168 B/op	  384486 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  83671371 ns/op	62784828 B/op	  384504 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      13	  90531701 ns/op	62752741 B/op	  384479 allocs/op
BenchmarkSiteNew/Edit_Deep_content_tree-8         	      12	  84060268 ns/op	62723508 B/op	  384494 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 110616452 ns/op	76831396 B/op	  450431 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 109269633 ns/op	76868226 B/op	  450432 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 105963661 ns/op	76884440 B/op	  450451 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 104363979 ns/op	76863216 B/op	  450428 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 105875770 ns/op	76948133 B/op	  450449 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 110638124 ns/op	76927885 B/op	  450446 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 110512414 ns/op	76875922 B/op	  450448 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 108193378 ns/op	76869274 B/op	  450440 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	      10	 105768744 ns/op	76926130 B/op	  450438 allocs/op
BenchmarkSiteNew/Regular_Deep_content_tree-8      	       9	 113038536 ns/op	76943137 B/op	  450433 allocs/op
PASS
ok  	github.com/gohugoio/hugo/hugolib	49.239s


name                                 old time/op    new time/op    delta
SiteNew/Edit_Deep_content_tree-8       77.6ms ± 9%    85.4ms ±17%   +9.97%  (p=0.006 n=9+10)
SiteNew/Regular_Deep_content_tree-8    94.2ms ± 5%   108.4ms ± 4%  +15.06%  (p=0.000 n=10+10)

name                                 old alloc/op   new alloc/op   delta
SiteNew/Edit_Deep_content_tree-8       50.2MB ± 0%    62.7MB ± 0%  +24.89%  (p=0.000 n=10+10)
SiteNew/Regular_Deep_content_tree-8    64.3MB ± 0%    76.9MB ± 0%  +19.54%  (p=0.000 n=10+10)

name                                 old allocs/op  new allocs/op  delta
SiteNew/Edit_Deep_content_tree-8         349k ± 0%      384k ± 0%  +10.01%  (p=0.000 n=10+10)
SiteNew/Regular_Deep_content_tree-8      415k ± 0%      450k ± 0%   +8.43%  (p=0.000 n=10+10)
```